Added the ability to create a DB of dependencies for each .i file
Runing dependies.exe creates a db of all of the files in a directory
passed to it, as well as the files that depend on that any specific file
to compile. On running, after the DB is built, the user can type the
name of a file in the directory passed, and a file
<filename.dependencies" will print. That "dependencies" file will hold
all of the files that need its parent to compile.